move thin wrappers to submodule level
