+++ src/myna/application/bnpy/cluster_supervoxel/__init__.py
@@ -1,0 +1,11 @@
+#
+# Copyright (c) Oak Ridge National Laboratory.
+#
+# This file is part of Myna. For details, see the top-level license
+# at https://github.com/ORNL-MDF/Myna/LICENSE.md.
+#
+# License: 3-clause BSD, see https://opensource.org/licenses/BSD-3-Clause.
+#
+"""Myna application for applying a clustering algorithm to supervoxelized data
+containing composition data, intended to be run after the bnpy/cluster_solidification
+app"""
