+++ src/myna/application/exaca/microstructure_region_slice/__init__.py
@@ -1,0 +1,11 @@
+#
+# Copyright (c) Oak Ridge National Laboratory.
+#
+# This file is part of Myna. For details, see the top-level license
+# at https://github.com/ORNL-MDF/Myna/LICENSE.md.
+#
+# License: 3-clause BSD, see https://opensource.org/licenses/BSD-3-Clause.
+#
+"""Myna application for calculating the microstructure statistics for a slice
+within a given region from reduced data format-type solidification data
+(FileGrainSlice)"""
